% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_rlda.R
\name{abc_rlda}
\alias{abc_rlda}
\title{Title}
\usage{
abc_rlda(X1, X2, H, S, c, gamma, kappa)
}
\arguments{
\item{X1}{asd}

\item{X2}{asd}

\item{H}{asd}

\item{S}{asd}

\item{c}{asd}

\item{gamma}{asd}

\item{kappa}{asd}
}
\value{
asd
}
\description{
Title
}
