% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{cross_validation}
\alias{cross_validation}
\title{Title}
\usage{
cross_validation(x, grouping, gamma, C_10, kfolds = 10)
}
\arguments{
\item{x}{Matrix or data.frame of observations.}

\item{grouping}{Grouping variable. A vector of numeric values 0 and 1 is recommended. Length has to correspond to nrow(x).}

\item{gamma}{regularization parameter}

\item{C_10}{parameter that controls prioretization of classes.
It's value should be between 0 and 1 (0 < cost_10 < 1)
Values bigger than 0.5 prioretizes correct classification of 0 class while values less than 0.5 prioretizes 1 class}
}
\value{
Returns average error of cross validation
}
\description{
Title
}
\examples{
range <- seq(0.05, 0.95, by=0.05)
gs <- grid_search(train, train_label,
                   range_gamma = grange,
                   range_C_10 = crange,
                   method = "estimator")
model <- abcrlda(train, train_label, gamma = gs0$gamma[1], cost_10 = gs0$C_10[1])
}
\seealso{
Other abcrlda binary classifier: \code{\link{abcrlda}},
  \code{\link{predict.abcrlda}}
}
\concept{abcrlda binary classifier}
