% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{grid_search}
\alias{grid_search}
\title{Grid Search}
\usage{
grid_search(x, y, range_gamma, range_cost, method = "estimator",
  nfold = 10)
}
\arguments{
\item{x}{Input matrix or data.frame , of dimension nobs x nvars; each row is an observation vector.}

\item{y}{Class labels. Should be a factor with two levels, or a vector with two distinct values.
If y is presented as a vector, it will be coerced into a factor.
Length of y has to correspond to number of samples in x.}

\item{range_gamma}{vector of gamma values to check}

\item{range_cost}{[1 x n] vector or [2 x n] matrix of cost values to check}

\item{method}{selects method to evaluete error. "estimator" and "cross"}

\item{nfold}{number of fold to use with cross-validation}
}
\value{
List of best founded parameters
}
\description{
Performs grid search based on cross validation or error estimation formula.
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                         iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                 iris[, ncol(iris)] == "versicolor"), 5])
cost_range <- seq(0.1, 0.9, by = 0.2)
gamma_range <- c(0.1, 1, 10, 100, 1000)

gs <- grid_search(train_data, train_label,
                  range_gamma = gamma_range,
                  range_cost = cost_range,
                  method = "estimator")
model <- abcrlda(train_data, train_label,
                 gamma = gs$gamma[1], cost = gs$cost[1])
predict(model, train_data)
}
\seealso{
Other Other functions in the package: \code{\link{abcrlda}},
  \code{\link{cross_validation}},
  \code{\link{da_risk_estimator}},
  \code{\link{predict.abcrlda}}
}
\concept{Other functions in the package}
