% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcrlda.R
\name{predict.abcrlda}
\alias{predict.abcrlda}
\title{Class Prediction for abcrlda objects}
\usage{
\method{predict}{abcrlda}(object, newx, out_type = "class", ...)
}
\arguments{
\item{object}{An object of class "abcrlda".}

\item{newx}{Matrix of new values for x at which predictions are to be made.}

\item{...}{Argument used by generic function predict(object, x, ...).}

\item{type}{Determines a type of output. Two type of input could be provided.
If "class" value is provided this will return factor with levels corresponding to lev stored in object.
If "raw" value is provided this will return numeric vector with values obtained from discriminant function.}
}
\value{
class Class prediction for each observation.
 raw Raw values.
}
\description{
Computes class predictions for new data based on a given abcrlda object
}
\section{Reference}{

  A. Zollanvari, M. Abdirash, A. Dadlani and B. Abibullaev,
  "Asymptotically Bias-Corrected Regularized Linear Discriminant Analysis for Cost-Sensitive
  Binary Classification," in IEEE Signal Processing Letters, vol. 26, no. 9, pp. 1300-1304,
  Sept. 2019. doi: 10.1109/LSP.2019.2918485
  URL: \url{http://ieeexplore.ieee.org/stamp/stamp.jsp?tp=&arnumber=8720003&isnumber=8770167}
}

\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                        iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                iris[, ncol(iris)] == "versicolor"), 5])
model <- abcrlda(train_data, train_label, gamma = 0.5, cost = 0.75)
predict(model, train_data)
}
\seealso{
Other Other functions in the package: \code{\link{abcrlda}},
  \code{\link{cross_validation}},
  \code{\link{da_risk_estimator}},
  \code{\link{grid_search}}
}
\concept{Other functions in the package}
