% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcrlda.R
\name{risk_calculate}
\alias{risk_calculate}
\title{Risk Calculate}
\usage{
risk_calculate(object, x_test, y_true)
}
\arguments{
\item{object}{An object of class "abcrlda".}

\item{x_test}{Matrix of values for x for which true class labels are known.}

\item{y_true}{a numeric vector or factor of true class labels. Factor should have two levels
or be a vector with two distinct values.
If \code{y_true} is presented as a vector, it will be coerced into a factor.
Length of \code{y_true} has to correspond to number of samples in \code{x_test}.}
}
\description{
Computes class predictions for new data based on a given abcrlda object
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                         iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                 iris[, ncol(iris)] == "versicolor"), 5])
model <- abcrlda(train_data, train_label, gamma = 0.5, cost = 0.75)
risk_calculate(model, train_data, train_label)
}
\seealso{
Other functions in the package: \code{\link{abcrlda}},
  \code{\link{cross_validation}},
  \code{\link{da_risk_estimator}},
  \code{\link{grid_search}}, \code{\link{predict.abcrlda}}
}
\concept{functions in the package}
