% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{risk_estimate_20}
\alias{risk_estimate_20}
\title{Title}
\usage{
risk_estimate_20(object)
}
\arguments{
\item{object}{An object of class "abcrlda".}
}
\value{
Estimated risk based on "abcrlda" object
}
\description{
Title
}
\examples{
range <- seq(0.05, 0.95, by=0.05)
gs <- grid_search(train, train_label,
                   range_gamma = grange,
                   range_C_10 = crange,
                   method = "estimator")
model <- abcrlda(train, train_label, gamma = gs0$gamma[1], cost_10 = gs0$C_10[1])
}
