% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcrlda.R
\name{abcrlda}
\alias{abcrlda}
\title{Asymptotically Bias-Corrected Regularized Linear Discriminant Analysis
for Cost-Sensitive Binary Classification}
\usage{
abcrlda(x, grouping, gamma, cost_10 = 0.5)
}
\arguments{
\item{x}{Matrix or data.frame of observations.}

\item{grouping}{Grouping variable. A vector of numeric values 0 and 1 is recommended.
Length has to correspond to nrow(x).}

\item{gamma}{regularization parameter}

\item{cost_10}{parameter that controls prioretization of classes.
It's value should be between 0 and 1 (0 < cost_10 < 1)
Values bigger than 0.5 prioretizes correct classification of 0 class while values
less than 0.5 prioretizes 1 class}
}
\value{
An object of class "rrlda" is returned which can be used for class prediction (see predict())
  \item{a }{Slope of a discriminant hyperplane. W(x) = a'x + m}
  \item{m }{Bias term. W(x) = a'x + m}
  \item{cost_10 }{asd}
  \item{gamma }{asd}
  \item{Ghat0}{asd}
  \item{Ghat1}{asd}
  \item{Dhat }{asd}
  \item{omegaopt }{asd}
  \item{lev }{Levels. Corresponds to the groups.}
}
\description{
Performs symptotically Bias-Corrected Regularized Linear Discriminant Analysis
}
\examples{
data(iris)
traindata = iris[ which(iris[,ncol(iris)]=='virginica' |
                        iris[,ncol(iris)]=="versicolor"), 1:4]
trainlabel = factor(iris[ which(iris[,ncol(iris)]=='virginica' |
                                iris[,ncol(iris)]=="versicolor"), 5])
rr <- abcrlda(traindata, trainlabel, gamma = 0.5, cost_10 = 0.75)
predict(rr, traindata)
}
\seealso{
Other abcrlda binary classifier: \code{\link{cross_validation}},
  \code{\link{predict.abcrlda}}
}
\concept{abcrlda binary classifier}
