% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation}
\usage{
cross_validation(x, grouping, gamma = 1, cost = c(0.5, 0.5),
  kfolds = 10)
}
\arguments{
\item{x}{Input matrix or data.frame , of dimension nobs x nvars; each row is an observation vector.}

\item{gamma}{regularization parameter}

\item{cost}{parameter that controls prioretization of classes.
It's value should be between 0 and 1 (0 < cost_10 < 1)
Values bigger than 0.5 prioretizes correct classification of 0 class while values less than 0.5 prioretizes 1 class}

\item{kfolds}{Number of for cross validation algorithm}
}
\value{
Returns average error of cross validation
}
\description{
Cross Validation
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                         iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                 iris[, ncol(iris)] == "versicolor"), 5])
model <- abcrlda(train_data, train_label, gamma = 0.5, cost = 0.75)
risk_estimate_20(model)
}
\seealso{
Other Other functions in the package: \code{\link{abcrlda}},
  \code{\link{da_risk_estimator}},
  \code{\link{grid_search}}, \code{\link{predict.abcrlda}}
}
\concept{Other functions in the package}
