% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation}
\usage{
cross_validation(x, grouping, gamma, C_10, kfolds = 10)
}
\arguments{
\item{x}{Matrix or data.frame of observations.}

\item{grouping}{Grouping variable. A vector of numeric values 0 and 1 is recommended. Length has to correspond to nrow(x).}

\item{gamma}{regularization parameter}

\item{C_10}{parameter that controls prioretization of classes.
It's value should be between 0 and 1 (0 < cost_10 < 1)
Values bigger than 0.5 prioretizes correct classification of 0 class while values less than 0.5 prioretizes 1 class}

\item{kfolds}{Number of for cross validation algorithm}
}
\value{
Returns average error of cross validation
}
\description{
Cross Validation
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                        iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                iris[, ncol(iris)] == "versicolor"), 5])
cost_range <- seq(0.1, 0.9, by = 0.2)
gamma_range <- c(0.1, 1, 10, 100, 1000)

gs <- grid_search(train_data, train_label,
                  range_gamma = gamma_range,
                  range_C_10 = cost_range,
                  method = "estimator")
model <- abcrlda(train_data, train_label,
                 gamma = gs$gamma[1], cost_10 = gs$C_10[1])
predict(model, train_data)
}
\seealso{
Other abcrlda binary classifier: \code{\link{abcrlda}},
  \code{\link{predict.abcrlda}}
}
\concept{abcrlda binary classifier}
