% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{da_risk_estimator}
\alias{da_risk_estimator}
\title{Double Asymptotic Risk Estimator}
\usage{
da_risk_estimator(object)
}
\arguments{
\item{object}{An object of class "abcrlda".}
}
\value{
Calculates risk based on estimated class error rates and misclassification costs
  \deqn{\Re = \varepsilon_0 * C_{10} + \varepsilon_1 * C_{01}}{R = e_0 * C_10 + e_1 * C_01)}
}
\description{
This function implements the generalized (double asymptotic) consistent estimator of risk.
}
\section{Reference}{

  A. Zollanvari, M. Abdirash, A. Dadlani and B. Abibullaev,
  "Asymptotically Bias-Corrected Regularized Linear Discriminant Analysis for Cost-Sensitive
  Binary Classification," in IEEE Signal Processing Letters, vol. 26, no. 9, pp. 1300-1304,
  Sept. 2019. doi: 10.1109/LSP.2019.2918485
  URL: \url{https://ieeexplore.ieee.org/document/8720003}
}

\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                         iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                 iris[, ncol(iris)] == "versicolor"), 5])
model <- abcrlda(train_data, train_label, gamma = 0.5, cost = 0.75)
da_risk_estimator(model)
}
\seealso{
Other functions in the package: 
\code{\link{abcrlda}()},
\code{\link{cross_validation}()},
\code{\link{grid_search}()},
\code{\link{predict.abcrlda}()},
\code{\link{risk_calculate}()}
}
\concept{functions in the package}
