% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{grid_search}
\alias{grid_search}
\title{Title}
\usage{
grid_search(x, grouping, range_gamma, range_C_10, method = "estimator",
  k_fold = 10)
}
\arguments{
\item{x}{Matrix or data.frame of observations.}

\item{grouping}{Grouping variable. A vector of numeric values 0 and 1 is recommended. Length has to correspond to nrow(x).}

\item{range_gamma}{vector of gamma values to check}

\item{range_C_10}{vector of cost values to check}

\item{method}{selects method to evaluete error. "estimator" and "cross"}

\item{k_fold}{number of fold to use with cross-validation}
}
\value{
List of best founded parameters
}
\description{
asd
}
\examples{
range <- seq(0.05, 0.95, by=0.05)
gs <- grid_search(train, train_label,
                   range_gamma = grange,
                   range_C_10 = crange,
                   method = "estimator")
model <- abcrlda(train, train_label, gamma = gs0$gamma[1], cost_10 = gs0$C_10[1])
}
