% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{grid_search}
\alias{grid_search}
\title{Grid Search}
\usage{
grid_search(x, grouping, range_gamma, range_cost, method = "estimator",
  k_fold = 10)
}
\arguments{
\item{x}{Matrix or data.frame of observations.}

\item{grouping}{Grouping variable. A vector of numeric values 0 and 1 is recommended. Length has to correspond to nrow(x).}

\item{range_gamma}{vector of gamma values to check}

\item{range_cost}{[1 x n] vector or [2 x n] matrix of cost values to check}

\item{method}{selects method to evaluete error. "estimator" and "cross"}

\item{k_fold}{number of fold to use with cross-validation}
}
\value{
List of best founded parameters
}
\description{
Performs grid search based on cross validation or error estimation formula.
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                         iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                 iris[, ncol(iris)] == "versicolor"), 5])
cost_range <- seq(0.1, 0.9, by = 0.2)
gamma_range <- c(0.1, 1, 10, 100, 1000)

gs <- grid_search(train_data, train_label,
                  range_gamma = gamma_range,
                  range_cost = cost_range,
                  method = "estimator")
model <- abcrlda(train_data, train_label,
                 gamma = gs$gamma[1], cost = gs$cost[1])
predict(model, train_data)
}
