% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcrlda.R
\name{predict.abcrlda}
\alias{predict.abcrlda}
\title{Class Prediction for abcrlda objects}
\usage{
\method{predict}{abcrlda}(object, x, type = "class", ...)
}
\arguments{
\item{object}{An object of class "abcrlda".}

\item{x}{New data for which the classes are to predict}

\item{type}{TODO}

\item{...}{Argument used by generic function predict(object, x, ...).}
}
\value{
class Class prediction for each observation.
 raw Raw values.
}
\description{
Computes class predictions for new data based on a given abcrlda object
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                        iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                iris[, ncol(iris)] == "versicolor"), 5])
model <- abcrlda(train_data, train_label, gamma = 0.5, cost = 0.75)
predict(model, train_data)
}
\seealso{
Other abcrlda binary classifier: \code{\link{abcrlda}},
  \code{\link{cross_validation}}
}
\concept{abcrlda binary classifier}
