% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_grid.R
\name{risk_estimate_20}
\alias{risk_estimate_20}
\title{Risk Estimator}
\usage{
risk_estimate_20(object)
}
\arguments{
\item{object}{An object of class "abcrlda".}
}
\value{
Weighted error based on "abcrlda" object
}
\description{
Calculates weighted error based on normalized cost values
}
\examples{
data(iris)
train_data <- iris[which(iris[, ncol(iris)] == "virginica" |
                         iris[, ncol(iris)] == "versicolor"), 1:4]
train_label <- factor(iris[which(iris[, ncol(iris)] == "virginica" |
                                 iris[, ncol(iris)] == "versicolor"), 5])
model <- abcrlda(train_data, train_label, gamma = 0.5, cost = 0.75)
risk_estimate_20(model)
}
